chore(rac): separate rac extensions in 2 separate files
